Add --nat flag
