+++ packages/beacon-node/src/network/nodejs/util.ts
@@ -4,7 +4,6 @@ import {ENR, SignableENR} from "@chainsafe/discv5";
 import {Libp2p} from "../interface.js";
 import {Eth2PeerDataStore} from "../peers/datastore.js";
 import {defaultDiscv5Options, defaultNetworkOptions, NetworkOptions} from "../options.js";
-import {isLocalMultiAddr, clearMultiaddrUDP} from "../util.js";
 import {createNodejsLibp2p as _createNodejsLibp2p} from "./bundle.js";
 
 export type NodeJsLibp2pOpts = {
@@ -31,14 +30,8 @@ export async function createNodeJsLibp2p(
   const enr = networkOpts.discv5?.enr;
   const {peerStoreDir, disablePeerDiscovery} = nodeJsLibp2pOpts;
 
-  if (enr !== undefined && typeof enr !== "string") {
-    if (enr instanceof SignableENR) {
-      if (enr.getLocationMultiaddr("udp") && !isLocalMultiAddr(enr.getLocationMultiaddr("udp"))) {
-        clearMultiaddrUDP(enr);
-      }
-    } else {
-      throw Error("network.discv5.enr must be an instance of ENR");
-    }
+  if (enr !== undefined && typeof enr !== "string" && !(enr instanceof SignableENR)) {
+    throw Error("network.discv5.enr must be an instance of SignableENR");
   }
 
   let datastore: undefined | Eth2PeerDataStore = undefined;
