+++ packages/beacon-node/test/unit/network/util.test.ts
@@ -1,24 +1,11 @@
-import {multiaddr} from "@multiformats/multiaddr";
 import {expect} from "chai";
 import {createSecp256k1PeerId} from "@libp2p/peer-id-factory";
 import {config} from "@lodestar/config/default";
 import {ForkName} from "@lodestar/params";
 import {generateKeypair, KeypairType, SignableENR} from "@chainsafe/discv5";
 import {defaultNetworkOptions} from "../../../src/network/options.js";
-import {createNodeJsLibp2p, isLocalMultiAddr} from "../../../src/network/index.js";
+import {createNodeJsLibp2p} from "../../../src/network/index.js";
 import {getCurrentAndNextFork} from "../../../src/network/forks.js";
-
-describe("Test isLocalMultiAddr", () => {
-  it("should return true for 127.0.0.1", () => {
-    const multi0 = multiaddr("/ip4/127.0.0.1/udp/30303");
-    expect(isLocalMultiAddr(multi0)).to.equal(true);
-  });
-
-  it("should return false for 0.0.0.0", () => {
-    const multi0 = multiaddr("/ip4/0.0.0.0/udp/30303");
-    expect(isLocalMultiAddr(multi0)).to.equal(false);
-  });
-});
 
 describe("getCurrentAndNextFork", function () {
   const altairEpoch = config.forks.altair.epoch;
